Added user-specified metadata keys
These keys will allow users to spawn multiple versions of the beat
that can export data into multiple locations without messing up each
others' processing (under default configuration).

Also included is a fix for the iterator in storageprovider that
wouldn't terminate when called in a for-loop (it's not a standard go
iterator).

Change-Id: I1ee9dffdfaf8f5c6b6ed2e4d777489d07f1004d0